refactor vertex material creation
