+++ tests/common/cases/test_utils.py
@@ -13,155 +13,12 @@ from tests.common.helpers.collision_box import *
 from tests.common.helpers.hierarchy import *
 from tests.common.helpers.hlod import *
 from tests.common.helpers.mesh import *
-from tests.common.helpers.mesh_structs.shader_material import *
 from tests.utils import *
 from tests.w3d.helpers.compressed_animation import *
 from tests.w3d.helpers.dazzle import *
-from tests.w3d.helpers.mesh_structs.shader import *
-from tests.w3d.helpers.mesh_structs.vertex_material import *
 
 
 class TestUtils(TestCase):
-    def test_vertex_material_roundtrip(self):
-        expected = get_vertex_material()
-
-        copyfile(up(up(self.relpath())) + '/testfiles/texture.dds', self.outpath() + 'texture.dds')
-
-        (material, _) = create_material_from_vertex_material(
-            self, 'meshName', expected)
-        actual = retrieve_vertex_material(material)
-        compare_vertex_materials(self, expected, actual)
-
-    def test_vertex_material_no_attributes_roundtrip(self):
-        expected = get_vertex_material()
-
-        copyfile(up(up(self.relpath())) + '/testfiles/texture.dds',
-                 self.outpath() + 'texture.dds')
-
-        expected.vm_info.attributes = 0
-        (material, _) = create_material_from_vertex_material(
-            self, 'meshName', expected)
-        actual = retrieve_vertex_material(material)
-        compare_vertex_materials(self, expected, actual)
-
-    def test_shader_material_roundtrip(self):
-        expected = get_shader_material()
-
-        copyfile(up(up(self.relpath())) + '/testfiles/texture.dds', self.outpath() + 'texture.dds')
-
-        (material, _) = create_material_from_shader_material(
-            self, 'meshName', expected)
-        principled = node_shader_utils.PrincipledBSDFWrapper(material, is_readonly=True)
-        actual = retrieve_shader_material(material, principled)
-        compare_shader_materials(self, expected, actual)
-
-    def test_duplicate_shader_material_roundtrip(self):
-        expecteds = [get_shader_material(), get_shader_material()]
-
-        materials = []
-        for mat in expecteds:
-            (material, _) = create_material_from_shader_material(self, 'meshName', mat)
-            materials.append(material)
-
-        self.assertEqual(1, len(bpy.data.materials))
-        self.assertTrue('meshName.NormalMapped.fx' in bpy.data.materials)
-
-        for expected in expecteds:
-            principled = node_shader_utils.PrincipledBSDFWrapper(material, is_readonly=True)
-            actual = retrieve_shader_material(self, material, principled)
-            compare_shader_materials(self, expected, actual)
-
-    def test_shader_material_w3x_roundtrip(self):
-        expected = get_shader_material(w3x=True)
-        copyfile(up(up(self.relpath())) + '/testfiles/texture.dds',
-                 self.outpath() + 'texture.dds')
-
-        (material, _) = create_material_from_shader_material(
-            self, 'meshName', expected)
-        principled = node_shader_utils.PrincipledBSDFWrapper(material, is_readonly=True)
-        actual = retrieve_shader_material(material, principled, w3x=True)
-        compare_shader_materials(self, expected, actual)
-
-    def test_shader_material_w3x_rgb_colors_roundtrip(self):
-        expected = get_shader_material(w3x=True, rgb_colors=True)
-
-        (material, _) = create_material_from_shader_material(
-            self, 'meshName', expected)
-        principled = node_shader_utils.PrincipledBSDFWrapper(material, is_readonly=True)
-        actual = retrieve_shader_material(self, material, principled, w3x=True)
-
-        for prop in expected.properties:
-            if prop.name in ['ColorAmbient', 'ColorEmissive', 'ColorDiffuse', 'ColorSpecular']:
-                prop.type = 5
-                prop.value = Vector((prop.value[0], prop.value[1], prop.value[2], 1.0))
-
-        compare_shader_materials(self, expected, actual)
-
-    def test_shader_material_w3x_two_tex_roundtrip(self):
-        expected = get_shader_material(w3x=True, two_tex=True)
-
-        (material, _) = create_material_from_shader_material(
-            self, 'meshName', expected)
-        principled = node_shader_utils.PrincipledBSDFWrapper(material, is_readonly=True)
-        actual = retrieve_shader_material(self, material, principled, w3x=True)
-        compare_shader_materials(self, expected, actual)
-
-    def test_default_shader_material_properties_are_not_exported(self):
-        expected = get_shader_material()
-        expected.properties = []
-
-        (material, principled) = create_material_from_shader_material(self, 'meshName', expected)
-
-        actual = retrieve_shader_material(self, material, principled, w3x=False)
-        self.assertEqual(0, len(actual.properties))
-
-        actual = retrieve_shader_material(self, material, principled, w3x=True)
-        self.assertEqual(0, len(actual.properties))
-
-    def test_shader_material_minimal_roundtrip(self):
-        mesh = get_mesh(shader_mats=True)
-
-        for source in mesh.shader_materials:
-            source.properties = get_shader_material_properties_minimal()
-
-            (material, principled) = create_material_from_shader_material(self, mesh.name(), source)
-            actual = retrieve_shader_material(self, material, principled)
-            source.properties[2].type = 5
-            source.properties[2].value = get_vec4(x=1.0, y=0.2, z=0.33, w=1.0)
-            compare_shader_materials(self, source, actual)
-
-    def test_shader_material_type_name_fallback(self):
-        mesh = get_mesh(shader_mats=True)
-
-        for source in mesh.shader_materials:
-            source.header.type_name = 'LoremIpsum'
-            source.properties = []
-
-            (material, principled) = create_material_from_shader_material(self, mesh.name(), source)
-            actual = retrieve_shader_material(self, material, principled)
-            source.header.type_name = 'DefaultW3D.fx'
-            compare_shader_materials(self, source, actual)
-
-    def test_shader_material_type_name_upgrade_to_normal_mapped(self):
-        mesh = get_mesh(shader_mats=True)
-
-        for source in mesh.shader_materials:
-            source.header.type_name = 'LoremIpsum'
-
-            (material, principled) = create_material_from_shader_material(self, mesh.name(), source)
-            actual = retrieve_shader_material(self, material, principled)
-            source.header.type_name = 'NormalMapped.fx'
-            compare_shader_materials(self, source, actual)
-
-    def test_shader_roundtrip(self):
-        mesh = get_mesh()
-
-        (material, _) = create_material_from_vertex_material(mesh.name(), mesh.vert_materials[0])
-        expected = mesh.shaders[0]
-        set_shader_properties(material, expected)
-        actual = retrieve_shader(material)
-        compare_shaders(self, expected, actual)
-
     def test_boxes_roundtrip(self):
         hlod = get_hlod()
         hlod.lod_arrays[0].sub_objects.append(get_hlod_sub_object(bone=1, name='containerName.WORLDBOX'))
